backend_fastapi/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
mongoengine>=0.27.0
pymongo>=4.6.0
bcrypt>=4.1.0
PyJWT>=2.8.0
python-multipart>=0.0.6
cloudinary>=1.36.0
